Completed Icons/Text color feature
its not pretty code, but because each icon is a pseudo ::after element, JavaScript is not letting me just use the class to change the style.color. I bet theres a better solution for this, but I want to allocate my time implementing more features. I don't think its going to such a problem considering its just a small project.